# Conflicts:
#	matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/room/relation/DefaultRelationService.kt

+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/file/FileService.kt
@@ -117,5 +117,5 @@ interface FileService {
     /**
      * Get size of cached files
      */
-    fun getCacheSize(): Int
+    fun getCacheSize(): Long
 }
